fixed getValueOf funct
